refactor: CreateMemberRequest 패키지 이동
