+++ src/main/java/cart/dto/request/CreateMemberRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.request;
 
 public class CreateMemberRequest {
